TOOL-8: Reorganize StreamRepository Tests
